ai-core/requirements.txt
sentence-transformers
faiss-cpu==1.10.0
watchdog
numpy
